# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/servicios": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener lista de servicios paginada con filtros dinámicos",
                "operationId": "3a1c1c392788bd4c8d03626d48d28645",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre del servicio para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Descripción del servicio para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Categoría del servicio para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de servicios filtrada y paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "cd978e7e6c3a84104d6e47bc1775f0f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServicioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener un servicio por ID",
                "operationId": "9c04868e405cd028d12617722c41d260",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del servicio"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Actualizar un servicio",
                "operationId": "44c627d011d44594d0978735c2198fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutServicioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servicio actualizado correctamente"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Eliminar un servicio",
                "operationId": "fdd7b675238376844570194b662ea1da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Actualizar un servicio",
                "operationId": "36358ec90e3e35793e3671f860d7e1ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutServicioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servicio actualizado correctamente"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            }
        },
        "/api/servicios/{id}/hoteles": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener hoteles asociados a un servicio",
                "operationId": "1dfb1676fd6f90729fcad5ade8eb9f69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hoteles asociados"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Permite a un usuario iniciar sesión con sus credenciales (email y contraseña). Si las credenciales son correctas, se devuelve un token de autenticación.",
                "operationId": "e2f16b490348a902bb9994972e2d20af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso, se devuelve el token de autenticación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMjM0NTY3ODkwLCJpYXQiOjE1MTYyMzkwMjJ9.dXnp_wHjoOFlvOzx4f56zuwFOjwKBGjvv3wbFsQdSuM"
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas (usuario o contraseña inválidos)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Usuario y/o contraseña invalida"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreServicioRequest": {
                "title": "Store Servicio Request",
                "description": "Datos requeridos para crear un nuevo servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PutServicioRequest": {
                "title": "Put Servicio Request",
                "description": "Datos requeridos para actualizar un servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}